Sort the changes to repos
